Fix: Another unnamed CacheManager already exists
When GeneServiceImplTest runs after any DataJpaTest this
error occurs. This is because DataJpaTests create an CacheManager
in order to allow for ehcache second level and query caching. The
GeneServiceImplTest is not aware of this CacheManager and attempts
to create its own. Multiple CacheManagers under he same name are
no longer allowed in EhCache 2.5+.

Fixes explored:
-Turn off second level caching and query caching for DataJpaTests.
  This works but is not optimal if we would like to test caching
  in the future.
-Apply @DirtiesContext to relevant Test classes. This method is
  very fragile, see https://jira.spring.io/browse/SPR-12687.
-Ignore the Spring managed CacheManager and create a standalone
  one under a different name for the purposes of
  GeneServiceImpleTest. This was ultimately the cleanest solution
  although it might not play well if further development combines
  Gene caching and hibernate caching.