Add initial work for gene-level privacy
Add a nullable gene_privacy_level column to the gene table to indicate
gene-specific privacy level.

Add PrivacyLevelType so that privacy levels can be annotated with labels
and descriptions.

Add PrivacySensitive interface to annotate models with an optional owner
and a privacy level.

Add a privacy service to centralize the manipulation of privacy levels.

Add two options to configure available levels and enable gene-level
privacy.