Defer Flyway initialization after Hibernate
We don't yet use Flyway to create the initial database, so we have to
make sure it does not run before Hibernate has created the tables.

Add a configuration and import it in repositories tests. This has the
side effect of systematically validating the migration scripts when test
are run.