Fix warnings from Intellij
Use Class::method instead of lambda wherever possible.

Initialize the template name in ModelAndView constructor instead of via
setViewName to ensure that the view is always valid.

Fix detected issue with invalid tier types not raising a bad request
error in the API and add appropriate tests.

Remove usage of Optional<> as argument parameter as it was intended for
return values.