Use timestamps for password and verification tokens (fix #38)
Use timestamps to store token expiration instead of datetime to describe
a precise moment in time that is tied to a UTC timezone.

Add a email in the VerificationToken to ensure that the email being
verified still match what is set in the user account.

Make user.enabled a non-nullable field.

Delete used tokens to prevent them from being reused.