search: Add user profile popover for remote users (fix #35 and #56)
Handle empty popover by raising a 204 No Content status.

Ensure that remote is >=1.4.0 to enable popover for anonymous remote
users.

Add @PreAuthorize in the remote resource service to check for
international search permission.

Rename findUserByAnonymousId and findUserGeneByAnonymousId to clearly
indicate that no permission check is performed on the returned model.