Make GeneOntologyTerm abstract and add GeneOntologyTermInfo
Using GeneOntologyTerm directly is problematic because it is not
interchangeable with UserTerm since the two classes have distinct
hashCode and equals definitions.

Instead, make it abstract and introduce a new class for representing
static GO term information.

Add a GeneOntologyTermInfoRepository that implements the CrudRepository
interface to manage in-memory persistence of GO terms and simplify the
service layer.