Add public users count and access to anonymized users and genes (fix #69)
Add a public users count to the /api/stats endpoint that quantifies the
number of public user profiles available in a registry.

Add two new endpoints for listing users and genes in a paginated format.
Users and genes that are not available are anonymized. There's no way to
establish the relationship between a user and a gene.

These endpoints are only available if anonymized search results are
enabled because there's no clean way to filter the result of a paginated
access at this moment.