Use natural IDs for hashCode and equals
To correctly use equals() and hashCode() in Hibernate, entities must
have either a non-generated primary key or a natural ID. The notion of
natural ID is a set of columns that can unique describe individual
entities. The main reason is that the state before/after persistence
must not change the return value of equals() nor hashCode().

Make 'email' the natural ID of a user class and fix a few tests with
flawed assumptions.

Taxon uses a non-generated NCBI IDs, so it can safely use it for that
purpose.

Move the unique constraint attribute from Token subclasses to the
@Column attribute of the token itself.

Remove a few redundant @Data since we already define @EqualsAndHashCode
as well as @ToString.

Make the role attribute unique so that it can be used as a natural ID.
This introduces a change in the DB.