Use afterPropertiesSet() to initialize GO terms at boot time and improve logging
Use String.format() for logging.

Report the amout of time it took to update GO terms.

Use a background thread to initialize GO terms until the scheduled
update kicks in. This means we must make updateGoTerms() synchronized to
avoid race condition between initialization and a scheduled update.

This reduces the boot time to approximately 7 seconds!