Unified user search (fix #203)
Fix findGenesBySymbol test with TIER 3 restriction, the tiers traversing
order was undefined, which violated the mock server expectations.

Consistently use the request timeout in the RemoteResourceService so
that assessing the API version does not impede getRemoteEntities().

Beef up SearchController tests that were missing a lot of verify()
statements.

Introduce a constant and a cached service utility to test if an ontology
is deemed simple.