Fixes for the ontology selection fragment
Fix incorrect display of icon due to a mix of th:if and th:replace (fix #248).

Align items vertically (and horizontally using mx-auto) when laying out
icons of different sizes. This only applies to the read-only view
because otherwise it would misalign the checkboxes.

Don't display the autocomplete input for a read-only table
presentation.t