Make GO term search more efficient
Avoid copying the internal state of the GO repository by providing a new
Stream-based API. It uses a read lock that is released when the stream
is closed.

Use a thread-local storage for the GO term size as it can be overwritten
by another thread processing a different autocomplete request.

Make extras in SearchResult non-final so we can add them after the fact.

Add a few logging to report slow search.