Warn users when terms are missing from partner registries (fix #300)
When performing a search with terms available locally, it can be a bit
puzzling to the user which criterias will be applied to partner
registries that might not have the same set of terms and categories
enabled.

Thus we introduce here a new endpoint for querying the availability of
multiple ontology terms and summarize that information in a warning when
applicable.

Introduce a new RemoteResource interface to hold the origin and
originUrl attributes.

Make RemoteResource and ApiController return remote flavours of Ontology
and OntologyTermInfo.