Fix a bunch of issues from SonarQube
Centralize the cache creation in a CacheConfig and inject relevant cache
beans in the services that need them.

Use an immutable and thread-safe DateTimeFormatter for formatting the
'date:' field in the OBO format. In addition, use the time zone of the
user requesting the download.

Sanitize a bunch of logs against newlines.