Fix ordering of FAQ entries
Using a LinkedHashMap is insufficient to guarantee the ordering of keys
as they are first inserted in a java.util.Properties mapping.

Instead, introduce a new `rdp.faq.keys` setting that specify the
ordering of the FAQ entries.

Add user documentation regarding that feature and also how to translate
these messages.